On a changé la méthode Update du GraphRepository pour permettre le mise à jour de la propriete avec EdgeProperty

+++ AppCV-Fiches/DAL_CV-Fiches/Models/Graph/Technologie.cs
@@ -16,6 +16,6 @@ namespace DAL_CV_Fiches.Models.Graph
         public CategorieDeTechnologie Categorie { get; set; }
 
         [EdgeProperty]
-        public double MoisDExperience { get; set; }
+        public int MoisDExperience { get; set; }
     }
 }
